refactor: 패키지명 변경
model에서 domain으로 패키지명 변경
