+++ docs/README.md
@@ -20,15 +20,15 @@
 - [ ] 당첨 로또 번호와 번호가 중복되면 에러가 발생한다.
 
 
-- Lotteries
-- [ ] 발행된 로또를 가진다.
-
-
 - Lottery
 - [ ] 6개의 로또번호를 가진다.
 - [ ] 6개의 로또번호에 중복이 있으면 에러가 발생한다.
 - [ ] 당첨 번호와 매치하는 로또번호가 몇 개인지 확인한다.
-- [ ] 보너스번호가 로또번호에 포함되어 있는지 확인한다. 
+- [ ] 보너스번호가 로또번호에 포함되어 있는지 확인한다.
+
+
+- Lotteries
+- [ ] 발행된 로또를 가진다.
 
 
 - WinningResult
