+++ src/main/kotlin/lotto/domain/BonusNumber.kt
@@ -1,3 +1,3 @@
-package lotto.model
+package lotto.domain
 
 class BonusNumber(number: Int) : LotteryNumber(number)
