+++ src/main/kotlin/lotto/domain/PurchaseAmount.kt
@@ -1,4 +1,4 @@
-package lotto.model
+package lotto.domain
 
 class PurchaseAmount(
     val amount: Int
