+++ src/test/kotlin/lotto/domain/BonusNumberTest.kt
@@ -1,4 +1,4 @@
-package lotto.model
+package lotto.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
